add test googleTranslatorTest() in ByteBustersGroupTest class
class PR_GoogleTranslatorTest has been deleted
